Improve apollo queries exports
