+++ assets/src/domain/eventEditor/services/apollo/queries/datetimes/index.ts
@@ -1,44 +1,11 @@
-import gql from 'graphql-tag';
+export * from './queries';
 
-export const DATETIME_ATTRIBUTES: any = gql`
-	fragment datetimeAttributes on EspressoDatetime {
-		id
-		dbId
-		capacity
-		description
-		endDate
-		isActive
-		isExpired
-		isPrimary
-		isSoldOut
-		isTrashed
-		isUpcoming
-		length
-		name
-		order
-		reserved
-		sold
-		status
-		startDate
-	}
-`;
+export { default as useDatetimes } from './useDatetimes';
 
-export const GET_DATETIME: any = gql`
-	query GET_DATETIME($id: ID!) {
-		datetime(id: $id) {
-			...datetimeAttributes
-		}
-	}
-	${DATETIME_ATTRIBUTES}
-`;
+export { default as useDatetimeIds } from './useDatetimeIds';
 
-export const GET_DATETIMES: any = gql`
-	query GET_DATETIMES($where: EspressoRootQueryDatetimesConnectionWhereArgs) {
-		espressoDatetimes(where: $where) {
-			nodes {
-				...datetimeAttributes
-			}
-		}
-	}
-	${DATETIME_ATTRIBUTES}
-`;
+export { default as useDatetimeItem } from './useDatetimeItem';
+
+export { default as useDatetimeQueryOptions } from './useDatetimeQueryOptions';
+
+export { default as useFetchDatetimes } from './useFetchDatetimes';
