+++ assets/src/domain/eventEditor/services/apollo/queries/index.ts
@@ -49,3 +49,11 @@ export const DEFAULT_PRICE_TYPE_LIST_DATA: PriceTypeEdge = {
 	nodes: [],
 	__typename: 'EspressoRootQueryPriceTypesConnection',
 };
+
+export * from './datetimes';
+
+export * from './tickets';
+
+export * from './prices';
+
+export * from './priceTypes';
